Create separate bundles for each form component. Component import path will be @enigmatry/entry-form/xxx
