+++ projects/entry-form/autocomplete/filter-with-autocomplete.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { SelectOption } from '../../../interfaces';
+import { SelectOption } from './select-configuration.interface';
 
 @Pipe({
   name: 'filterWithAutocomplete'
