+++ projects/entry-form/autocomplete/formly-autocomplete.module.ts
@@ -5,10 +5,10 @@ import { FormlyModule } from '@ngx-formly/core';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatInputModule } from '@angular/material/input';
 import { FormlySelectModule } from '@ngx-formly/core/select';
+import { CheckAutocompleteInputDirective } from './check-autocomplete-input.directive';
+import { DisplayWithAutocompletePipe } from './display-with-autocomplete.pipe';
+import { FilterWithAutocompletePipe } from './filter-with-autocomplete.pipe';
 import { FormlyAutocompleteComponent } from './formly-autocomplete.component';
-import { CheckAutocompleteInputDirective } from './directives/check-autocomplete-input.directive';
-import { DisplayWithAutocompletePipe } from './pipes/display-with-autocomplete.pipe';
-import { FilterWithAutocompletePipe } from './pipes/filter-with-autocomplete.pipe';
 
 @NgModule({
   declarations: [
