+++ projects/entry-form/autocomplete/public-api.ts
@@ -1,0 +1,3 @@
+
+export { FormlyAutocompleteModule } from './formly-autocomplete.module';
+export * from './select-configuration.interface';
