+++ projects/entry-form/autocomplete/select-configuration.interface.ts
