+++ projects/entry-form/ckeditor/formly-ckeditor.component.html
