+++ projects/entry-form/ckeditor/ng-package.json
@@ -1,0 +1,5 @@
+{
+  "lib": {
+    "entryFile": "./public-api.ts"
+  }
+}
