+++ projects/entry-form/ckeditor/public-api.ts
@@ -1,0 +1,3 @@
+
+export { FormlyCkeditorModule } from './formly-ckeditor.module';
+export * from './ckeditor-options';
