+++ projects/entry-form/date-time-picker/date-fns-adapter/public-api.ts
@@ -1,0 +1,2 @@
+
+export { NgxMatDateFnsAdapter } from './date-fns-adapter';
