+++ projects/entry-form/date-time-picker/ng-package.json
@@ -1,0 +1,5 @@
+{
+  "lib": {
+    "entryFile": "./public-api.ts"
+  }
+}
