+++ projects/entry-form/date-time-picker/public-api.ts
@@ -1,0 +1,3 @@
+
+export { FormlyDateTimePickerModule } from './formly-date-time-picker.module';
+export * from './date-time-formats';
