+++ projects/entry-form/entry-form.module.ts
@@ -1,12 +1,10 @@
 import { NgModule } from '@angular/core';
 import { ENTRY_FIELD_TYPE_RESOLVER, fieldTypeResolverFactory, FieldTypeResolverService } from './services';
-import { StandardEntryComponentsModule } from './components/standard-entry-components.module';
 
 @NgModule({
   declarations: [
   ],
   imports: [
-    StandardEntryComponentsModule
   ],
   providers: [
     {
