+++ projects/entry-form/interfaces/index.ts
@@ -1,3 +1,3 @@
 export * from './field-expression-dictionary.interface';
-export * from './select-configuration.interface';
+export * from '../autocomplete/select-configuration.interface';
 export * from './form-config';
