+++ projects/entry-form/ng-package.json
@@ -2,9 +2,9 @@
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../dist/enigmatry/entry-form",
   "assets": [
-    "src/styles/**/*.scss"
+    "./styles/**/*.scss"
   ],
   "lib": {
-    "entryFile": "src/public-api.ts"
+    "entryFile": "./public-api.ts"
   }
 }
