+++ projects/entry-form/package.json
@@ -13,11 +13,10 @@
     "@angular/common": ">=13.0.0",
     "@angular/core": ">=13.0.0",
     "@ngx-formly/core": "^5.12.0",
-    "@ngx-formly/material": "^5.12.0",
-    "@ckeditor/ckeditor5-angular": "^4.0.0"
+    "@ngx-formly/material": "^5.12.0"
   },
   "scripts": {
-    "compile-scss": "sass --style=compressed --load-path=styles src/styles/partials/theming.scss src/styles/dist/main.css",
+    "compile-scss": "sass --style=compressed --load-path=styles styles/partials/theming.scss styles/dist/main.css",
     "lint": "stylelint --fix styles/**/*.scss"
   },
   "dependencies": {
