+++ projects/entry-form/public-api.ts
@@ -1,0 +1,7 @@
+/*
+ * Public API Surface of entry-form
+ */
+export * from './interfaces';
+export * from './services';
+
+export { EntryFormModule } from './entry-form.module';
