--- projects/entry-form/src/lib/components/standard-entry-components.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { FormlyAutocompleteModule } from './autocomplete/formly-autocomplete.module';
-
-@NgModule({
-  declarations: [],
-  exports: [
-    FormlyAutocompleteModule
-  ]
-})
-export class StandardEntryComponentsModule { }
