+++ projects/entry-form/styles/modules/_generator.scss
