+++ projects/entry-form/styles/partials/core/components/forms/_general.scss
